requirements.txt
langchain
faiss-cpu
sentence-transformers
openai
fastapi
uvicorn
python-dotenv
langdetect
pytest
